- Cover Controller and Views fully updated, - Category model created - CategoryId column added to genre - Still issue with Foreign key add between Genre and Category tables exists. Needs to avoid Genre table drop because it is actually filled with some data.

+++ Areas/Admin/Controllers/CoverController.cs
@@ -50,43 +50,63 @@ namespace Books.Areas.Admin.Controllers
             return View(cover);
         }
 
-        // GET: Admin/Cover/Edit/5
+        // GET: Admin/Cover/Edit
+        // -------------------------
         public async Task<IActionResult> Edit(short? id)
         {
+            // Check if id was properly sent
+            // -------------------------
             if (id == null)
             {
                 return NotFound();
             }
 
+            // If Id was properly sent then get Publisher from Db
+            // -------------------------
             var cover = await _db.Cover.FindAsync(id);
+
+            // If Publisher with such Id wasn't found then show Not Found screen
+            // -------------------------
             if (cover == null)
             {
                 return NotFound();
             }
+
+            // If all commands passed then display proper View
+            // -------------------------
             return View(cover);
         }
 
-        // POST: Admin/Cover/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Admin/Cover/Edit
+        // -------------------------
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(short id, [Bind("Id,Name")] Cover cover)
         {
+            // Validate id number
+            // -------------------------
             if (id != cover.Id)
             {
                 return NotFound();
             }
 
+            // Validate model state
+            // -------------------------
             if (ModelState.IsValid)
             {
+                // Try to update Cover name
+                // -------------------------
                 try
                 {
                     _db.Update(cover);
                     await _db.SaveChangesAsync();
                 }
+                // If cover update fails then throw an error page
+                // -------------------------
                 catch (DbUpdateConcurrencyException)
                 {
+                    // Check if selected cover exists in database
+                    // -------------------------
                     if (!CoverExists(cover.Id))
                     {
                         return NotFound();
@@ -96,37 +116,76 @@ namespace Books.Areas.Admin.Controllers
                         throw;
                     }
                 }
+
+                // Redirect to action page
+                // -------------------------
                 return RedirectToAction(nameof(Index));
             }
+
+            // If model state is invalid go back to Edit page
+            // -------------------------
             return View(cover);
         }
 
-        // GET: Admin/Cover/Delete/5
+        // GET: Admin/Cover/Delete
+        // -------------------------
         public async Task<IActionResult> Delete(short? id)
         {
+            // Check if id was properly sent
+            // -------------------------
             if (id == null)
             {
                 return NotFound();
             }
 
+            // If Id was properly sent then get Cover from Db
+            // -------------------------
             var cover = await _db.Cover
                 .FirstOrDefaultAsync(m => m.Id == id);
+
+            // If Cover with such Id wasn't found then show Not Found screen
+            // -------------------------
             if (cover == null)
             {
                 return NotFound();
             }
 
+            // If all commands passed then display proper View
+            // -------------------------
             return View(cover);
         }
 
-        // POST: Admin/Cover/Delete/5
+        // POST: Admin/Cover/Delete
+        // -------------------------
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(short id)
+        public async Task<IActionResult> DeleteConfirmed(short? id)
         {
+            // Check if id was properly sent
+            // -------------------------
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            // Search for cover with specified id
+            // -------------------------
             var cover = await _db.Cover.FindAsync(id);
+
+            // Check if cover exists in database
+            // -------------------------
+            if (cover == null)
+            {
+                return NotFound();
+            }
+
+            // If cover exists then remove it from database
+            // -------------------------
             _db.Cover.Remove(cover);
             await _db.SaveChangesAsync();
+
+            // Go back into index action
+            // -------------------------
             return RedirectToAction(nameof(Index));
         }
